requirements.txt
ansicolors==1.1.8
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
configparser==5.0.2
cryptography==3.4.6
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-crispy-forms==1.11.1
django-gulp==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
env-tools==2.2.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
Pillow==8.1.2
psutil==5.8.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
tini==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
